almost failed, something wrong, not sure if “int16_t” is a sufficient variable for bestMove, try without changing it ---------------------------------------------------- Elo   | -19.32 +- 18.41 (95%) SPRT  | 8.0+0.08s Threads=1 Hash=8MB LLR   | -2.35 (-2.94, 2.94) [0.00, 10.00] Games | N: 1044 W: 352 L: 410 D: 282 Penta | [87, 106, 165, 106, 58] https://programcidusunur.pythonanywhere.com/test/20/ ----------------------------------------------------

+++ src/structs.h
@@ -51,7 +51,7 @@ typedef struct {
     int16_t depth;       // current search depth
     uint8_t flag;       // flag the type of node (fail-high(score >= beta)/fail-low(score < alpha))
     int16_t score;       // score (alpha/beta/PV)
-    int16_t bestMove;
+    int bestMove;
 } tt;                 // transposition table (TT aka hash table)
